Enhance alignment configuration and documentation
- Added new CLI options and configuration parameters for `constrain_end_clip_to_start_offset` and `high_rate_recommended_refusion`, allowing users to control behavior when Chromaprint offsets disagree.
- Updated README.md to include descriptions of the new options and their effects on audio alignment processes.
- Refactored alignment logic to incorporate the new configuration settings, improving flexibility in handling offset discrepancies.
- Enhanced tests to validate the new configuration options and ensure correct functionality in various alignment scenarios.

These changes improve user control over audio alignment behavior and enhance the overall robustness of the alignment process.

+++ crates/clip-sync/src/application/offset_refinement.rs
@@ -808,7 +808,7 @@ mod tests {
                 .map(|i| {
                     let t = i as f32 / sample_rate as f32;
                     let freq = 300.0 + 400.0 * t;
-                    ((TAU as f32 * freq * t).sin() * (i16::MAX as f32 * 0.5)).round() as i16
+                    ((TAU * freq * t).sin() * (i16::MAX as f32 * 0.5)).round() as i16
                 })
                 .collect();
             let mut right = vec![0i16; count];
